feat(emergency): add background section in emergency pages
