+++ app/src/components/domain/CountryPastEventsChart/styles.module.css
@@ -1,4 +1,4 @@
-.past-events-chart {
+.country-past-events-chart {
     .data-point {
         color: var(--go-ui-color-primary-red);
     }
